(int) port

+++ Workerman/Connection/TcpConnection.php
@@ -233,7 +233,7 @@ class TcpConnection extends ConnectionInterface
     {
         if(!$this->_remotePort)
         {
-            if($this->_remoteAddress = stream_socket_get_name($this->_socket, true))
+            if($this->_remoteAddress = (int)stream_socket_get_name($this->_socket, true))
             {
                 list($this->_remoteIp, $this->_remotePort) = explode(':', $this->_remoteAddress, 2);
             }
